+++ clang/lib/Format/FormatTokenLexer.cpp
@@ -1633,12 +1633,12 @@ void FormatTokenLexer::readRawToken(FormatToken &Tok) {
   if ((Style.isJavaScript() || Style.isProto()) && Tok.is(tok::char_constant))
     Tok.Tok.setKind(tok::string_literal);
 
-  if (Tok.is(tok::comment) && isClangFormatOn(Tok.TokenText))
+  if (Tok.is(tok::comment) && isClangFormatOn(Tok.TokenText, Style))
     FormattingDisabled = false;
 
   Tok.Finalized = FormattingDisabled;
 
-  if (Tok.is(tok::comment) && isClangFormatOff(Tok.TokenText))
+  if (Tok.is(tok::comment) && isClangFormatOff(Tok.TokenText, Style))
     FormattingDisabled = true;
 }