+++ clang/lib/Format/IntegerLiteralSeparatorFixer.cpp
@@ -108,9 +108,9 @@ IntegerLiteralSeparatorFixer::process(const Environment &Env,
     auto Location = Tok.getLocation();
     auto Text = StringRef(SourceMgr.getCharacterData(Location), Length);
     if (Tok.is(tok::comment)) {
-      if (isClangFormatOff(Text))
+      if (isClangFormatOff(Text, Style))
         Skip = true;
-      else if (isClangFormatOn(Text))
+      else if (isClangFormatOn(Text, Style))
         Skip = false;
       continue;
     }