+++ clang/lib/Format/NumericLiteralCaseFixer.cpp
@@ -150,9 +150,9 @@ NumericLiteralCaseFixer::process(const Environment &Env,
     auto Location = Tok.getLocation();
     auto Text = StringRef(SourceMgr.getCharacterData(Location), Length);
     if (Tok.is(tok::comment)) {
-      if (isClangFormatOff(Text))
+      if (isClangFormatOff(Text, Style))
         Skip = true;
-      else if (isClangFormatOn(Text))
+      else if (isClangFormatOn(Text, Style))
         Skip = false;
       continue;
     }